+++ simulator/simulator.go
@@ -4,6 +4,8 @@ import (
 	"time"
 
 	"github.com/uhppoted/uhppote-core/messages"
+	"github.com/uhppoted/uhppote-core/types"
+
 	"github.com/uhppoted/uhppote-simulator/entities"
 )
 
@@ -22,4 +24,5 @@ type Simulator interface {
 	Open(door uint8) (bool, error)
 	Close(door uint8) (bool, error)
 	ButtonPressed(door uint8, duration time.Duration) (bool, error)
+	StoreCard(card uint32, from types.Date, to types.Date, doors []uint8, PIN uint32) error
 }